tests/wpt/web-platform-tests/css/css-break/parsing/orphans-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: parsing orphans with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-orphans">
<meta name="assert" content="orphans supports only the grammar '<integer>'.">
<meta name="assert" content="Negative values and zero are invalid.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("orphans", "auto");
test_invalid_value("orphans", "1 234");
test_invalid_value("orphans", "-234");
test_invalid_value("orphans", "-1");
test_invalid_value("orphans", "0");
</script>
</body>
</html>
